refactor: update EditDialog import path
